+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.8.0] - Not released
+### Added
+- Allow moderator to disable/enable invitations for specific user. Methods are:
+  - `POST /users/:username/disable-invites`
+  - `POST /users/:username/enable-invites`
 
 ## [2.7.0] - 2023-01-19
 ### Added